requirements.txt
streamlit==1.22.0
Flask==2.3.2
Flask-CORS==3.0.10
Pillow==9.5.0
requests==2.30.0
pyngrok==6.0.0
transformers==4.35.2
accelerate==0.25.0
bitsandbytes==0.41.1
torch>=2.0.0
torchvision
blinker==1.6.2
gTTS==2.3.2
pyttsx3==2.90
pygame==2.5.2
